// packages/medusa-client/src/admin/models/AdminGetDiscountParams.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AdminGetDiscountParams = {
  /**
   * Comma separated list of relations to include in the results.
   */
  expand?: string;
  /**
   * Comma separated list of fields to include in the results.
   */
  fields?: string;
};